nit: changes of file structure
